fix build error keymap helix:five_rows
Running `make helix:five_rows` was failing with an error.
The cause is the following part in render_logo() in file helix/rev2/keymaps/five_rows/oled_display.c.

```c
    oled_write_P(helix_logo, false);
    char buf[30];
    char *bufp;
```

Then I created the file `users/mtei/pseudo_sprintf.c` to make the contents of `helix/rev2/keymaps/five_rows/oled_display.c` and `helix/rev3_5rows/keymaps/five_rows/oled_display.c` the same.
