+++ users/mtei/pseudo_sprintf.h
@@ -1,0 +1,8 @@
+// Copyright 2022 Takeshi Ishii (@mtei)
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+
+char *sprints(char *buf, char *src);
+char *sprintd(char *buf, char *leadstr, int data);
+char *sprint2d(char *buf, char *leadstr, int data);
